// WebSystem/WebSystem/ClientApp/src/models/Formas/CNT/dasbal.ts
export interface IDasbal {
    id?: number;
    activo: string;
    pasivo: string;
    capital: string;
    egresos: string;
    ingresos: string;
    ord_activo: string;
    ord_pasivo: string;
    resultadod: string;
    resultadoa: string;
    acumuladod: string;
    acumuladoa: string;
    eg_cencos: boolean;
    in_cencos: boolean;
    caja: string;
    ventas: string;
    cto_venta: string;
    gasto_des: string;
    trasnpor: string;
    impuesto: string;
    inter_por: string;
}